Restructure core functions, add placeholder for config
