+++ lua/gitportal/config.lua
@@ -1,0 +1,13 @@
+
+local M = {}
+
+
+-- Default configuration
+local default = {
+  always_include_current_line = false
+}
+
+
+function M.setup(options)
+
+end
