+++ lua/gitportal/init.lua
@@ -1,1 +1,29 @@
-return require("gitportal.core")
+local cli = require('gitportal.cli')
+local git_helpers = require('gitportal.git')
+local config = require('yourplugin.config')
+
+local M = {}
+
+-- SETUP
+function M.setup(options)
+  config.setup(options)
+end
+
+-- CORE FUNCTIONS
+function M.open_file_in_browser()
+  local git_url = git_helpers.get_git_url_for_current_file()
+  if git_url ~= nil then
+    cli.open_link_in_browser(git_url)
+  end
+end
+
+
+function M.open_file_in_neovim()
+  local url = vim.fn.input("Github link > ")
+  if url ~= nil then
+    git_helpers.open_file_from_git_url(url)
+  end
+end
+
+
+return M
